Implement dropout for slot-specific values
Summary:
This diff tackles the problem of RNNG overfitting on slot-specific values. For example, "Call Mom" is a common utterance, and as a result, RNNG may overfit on the "Mom" token for other utterances to predict annotations scuh as ```[IN:CREATE_CALL call [SL:CONTACT Mom]]```.

Here, dropout is applied to a token embedding (before concatenation with the dictionary features) if the last non-shift, non-reduce oracle (gold) action was a slot (and not an intent).

Differential Revision: D15160322

fbshipit-source-id: c93ad52b0629254a74147bf7a7041c07ba9bb8a3