pytext - add metric reporter columns to schema (#714)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/714

currently, the column which metric reporter is looking at is not added to the schema.
This causes the code to fail, if the metric reporter is configed to look at a column for which no tensorizer is using.
This is needed when running OSS XLM on pre-processed data, but wanting to output raw text, for example for human evaluation of results.

Reviewed By: mwu1993

Differential Revision: D15908650

fbshipit-source-id: 50b257dcf727ac28c212721bfa16405882ab55d7