Adding weight init (#723)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/723

Adding weight initialization to the DeepCNN layer. Results suggest that the weight initialization does not have an effect on training stability or convergence. The experiment utilizing norm_clip and weight_init was either stopped before it overfit or it avoided it due to the clipping and the initialization. Comparisons over the experiments on frame level accuracy and both training and evaluation loss are below.

Frame Level Accuracy
{F162974260}

Evaluation Loss
{F162974297}

Training Loss
{F162974338}

Differential Revision: D15927540

fbshipit-source-id: 521bf7074da83afc091fc9bbf489a02adea0e6c3