Remove token & byte token assertion (#724)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/724

The token and byte token features in `ByteTokensDocumentModel` need to be the same length as they are concatenated, but we don't have to assert that the tokenizer configs are identical. It can be convenient to have different capitalizations of the features (lower-cased word embeddings, cased byte features), or a different append dialect option in multilingual mode). There already is a runtime check in `arrange_model_inputs`, so a failure in different tokenization will still happen reasonably early.

Reviewed By: chenyangyu1988

Differential Revision: D15975189

fbshipit-source-id: f3e390d6ab4c571f29599e89c60045edddb10d59