Create AlternatingRandomizedBatchSampler
Summary:
The AlternatingRandomizedBatchSampler is a RandomizedBatchSampler that alternates creating batches from two sets of keys. It takes two dictionaries of unnormalized iterator probabilities and chooses a batch using a key from the first dictionary using the corresponding probabilities and then a batch from the other dictionary the next round.

The AlternatingRandomizedBatchSampler is intended to be used during XLM pre-training to enforce a pattern of MLM, TLM, MLM, TLM tasks.

Differential Revision: D15931223

fbshipit-source-id: 137641002d175c5b0d7cdada28b0f3536b8bb78b