Create AlternatingRandomizedBatchSampler (#737)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/737

The AlternatingRandomizedBatchSampler is a RandomizedBatchSampler that alternates creating batches from two sets of keys. It takes two dictionaries of unnormalized iterator probabilities and chooses a batch using a key from the first dictionary using the corresponding probabilities and then a batch from the other dictionary the next round.

The AlternatingRandomizedBatchSampler is intended to be used during XLM pre-training to enforce a pattern of MLM, TLM, MLM, TLM tasks.

Reviewed By: rutyrinott

Differential Revision: D15931223

fbshipit-source-id: a2fa00815fae703c4e96c375c82090cbc706a2fc