knowledge distillation for fluent2 pytext (#730)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/730

Add knowledge distillation support to fluent2 PyTextDocumentClassifier transformer.

Look at the unit test added for an example of using the new behavior. At a high level, we add a "teacher_model_config" parameter which takes an optional DocModel.Config. If this is present, we train this model first, then use it to compute logits on the training/eval sets, and then train the final model based on these values with a SoftLabelTensorizer. We require the model_config to configure a loss function which is either KLDivergenceCELoss or KLDivergenceBCELoss.

Reviewed By: rutyrinott

Differential Revision: D15999729

fbshipit-source-id: 01196ea77fc40da78d2c7d3b59413015fb75aaf6