pytext distributed training in fluent2 (#743)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/743

Enable distributed training of PyText models in fluent2 via torch.multiprocessing.spawn.

This was a bit awkward to do as serializing tasks for sending to torch.multiprocessing.spawn isn't trivial. The easiest way to do it is to pass through the config objects directly, as they can pickle well.

In the future we should investigate making this easier so we can avoid the config interface; it looks like the third-party `dill` library would be good to investigate, as some simple testing showed that it could already serialize most of our tasks/models.

Reviewed By: rutyrinott

Differential Revision: D16015676

fbshipit-source-id: 506cf710525603843156f5cc5607e96269b8be0e