Change the create_optimizer()'s input parameters
Summary:
I changed the input of create_optimizer() from model to model_params in create_optimizer().
Since when creating the new optimizer wrapper, we need two versions of model_params.
If we cannot separate model and model params when creating the optimizer, we need to copy a model instead, which might be memory-inefficient.

Differential Revision: D16134253

fbshipit-source-id: 759d49017ee025ac75b573c276ddcdffd01c3e88