adapt prediction workflow to new design (#746)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/746

the prediction workflow only supports the old design, migrate it to support the new design. since I found some recurring jobs set using the old design, I didn't completely remove the old design support, but making it rather separate so it's easy to get rid of the old design in the future.

Reviewed By: jingfeidu

Differential Revision: D16096788

fbshipit-source-id: 7e51a2eebb86cda6452ecd0bd2405eceac989f4d