implemented lm scores reporting (#773)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/773

Implements scores reporting for the language model metric reporter. In this context, the score is the perplexity for each sentence in a batch. Additionally, scores aggregation can be memory intensive for large datasets, so there is a new option in the config to control the aggregation of metrics. If disabled, there will be no score reporting in the test_out file.

Differential Revision: D16207604

fbshipit-source-id: 322e0864e9310ea2367faae5f93d237dadc316ae