Move all token vocab creation to TokenTensorizer (#703)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/703

for the common models with word embeddings, the vocab for tokens is usually created in the `TokenTensorizer`'s initialize() method, then overwritten or added to during initialization of `WordEmbedding`. It's unexpected that a model's module initialization should alter the tensorizer's state, so vocab creation will now happen entirely in `TokenTensorizer` in two steps according to the flags:

- `build_from_data` - if True tokens from training data are added to vocab
  - `size_from_data` - maximum number of tokens to add from data

- `vocab_files` - if provided tokens from these files are added to vocab
  - `skip_header_line`, `lowercase_tokens`, `size_limit` modify how to read each vocab file

If vocab is already provided to the tensorizer (e.g. the `BERTTensorizer` subclass), the above steps are skipped.

Pretrained word embeddings will be initialized as before - for each token in the final vocab, we try to look it up in the pretrained word embedding file and use the embedding if found.

Reviewed By: seayoung1112

Differential Revision: D15835417

fbshipit-source-id: 4d9c0acad36f837d485711ecc56f03bd98394aa6