implemented perplexity reductions for lm score reporting (#779)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/779

Implements perplexity reductions for language model score reporting. Currently, the reported score is the average perplexity of all words in a sentence, but this may not work for all use cases (e.g., the presence of named entities in a sentence brings up the average perplexity, so the resulting metric may not be as useful). To this end, this diff introduces several ways to "reduce" the perplexity scores of a sentence: min, max, mean, median, and eos (i.e., the perplexity of the `__END_OF_SENTENCE__` token).

Differential Revision: D16244881

fbshipit-source-id: cd9126850a616a53c7f4cb241313764790b1222a