removed bilstm dependence on seq_lengths (#776)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/776

If seq_lengths is not provided as a parameter to BiLSTM and pack_sequence is disabled, then the BiLSTM representation should not use seq_lengths to initialize anything. This diff removes the dependence on seq_lengths in said scenarios by using the batch dimension of embedded_tokens to initialize zero hidden state.

This seemingly minor change makes a dramatic difference when exporting models to caffe2 because seq_lengths will no longer be a part of the model graph. This can give both latency and memory savings during inference, albeit small.

Reviewed By: arbabu123

Differential Revision: D16229675

fbshipit-source-id: 03a816006197b0c9f7d18ff139882506b998008c