edited target names during caffe2 export (#780)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/780

The caffe2 predict_net can't be run more than once if a) a vocabulary item has the same name as a caffe2 cell and b) the target names (i.e. vocabulary) is exported along with the model.

For an example, let's say our vocabulary contains the word "10" and there is also a cell 10 in the caffe2 graph not associated with the vocab word. When the caffe2 model generates predictions, it will write word 10's score to its associated cell, which also happens to be called cell 10. Because of this, the original cell 10's contents will be overwritten, causing issues in the next run of the predict_net.

This diff introduces a fix by appending an underscore to the beginning of each vocabulary item, ensuring that its name will not conflict with existing cell names in the graph.

Reviewed By: seayoung1112

Differential Revision: D16257838

fbshipit-source-id: 96ab49e24bb3e87e9b029a43094c0a01a39d0100