Migrate create_predictor() to new data API (#781)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/781

issue in Github https://github.com/facebookresearch/pytext/issues/701
root cause: create_predictor() is broken due to "features" is removed in the new Data API.
We generate a predictor by create_predictor() and feed the predictor with 1 row of data(typically raw_text in json) in each call. However, in the new API, we feed predictor with a batch of data. To migrate it to new API, We can either 1) construct a special single row batch and follow https://fburl.com/tp6ekiqa to get model_inputs or 2) generate the model_inputs in adhoc way

As I don't see an internal use case to predict with 1 row of data and approach 1) will take much efforts to implement(need to create a new Batcher to read data from a row of json input), I'm doing 2) in this diff.

Reviewed By: Titousensei

Differential Revision: D16268791

fbshipit-source-id: 5a357773410162ab0bf986f681413ae14c09dde8