Add Recall at Precision Thresholds to Config (#792)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/792

Currently, recall at precision thresholds are fixed at the defaults of 0.2, 0.4, 0.6, 0.8, 0.9. Some use cases are concerned with R@P95 or other thresholds, so it would be more convenient to have these thresholds specifiable in the config of PyText flow runs.

Adds recall_at_precision_thresholds to ClassificationMetricReporter (and MultiLabelClassificationMetricReporter) as a list of floats.

Differential Revision: D16346078

fbshipit-source-id: 7d569b188c798cdfff3f2138f585187e0b0c76ea