improve csv support in TSVDataSource (#777)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/777

TSVDataSource documentation for delimiter param says Change to "," for csv,
but csv files often have quoted fields and TSVDataSource does not support
quoted fields.

We cannot blindly force all fields starting with quotes to be treated as
quoted fields, because it drastically changes the behavior: unclosed quoted
fields will merge with the next row, swallowing \n characters until we find
the closing quote. Some data sets might contain unclosed fields with quotes
and rely on the current behavior.

This diff adds a parameter to the TSVDataSource config that allows users to
specify whether they want quoted fields. The default is False, which is the
current behavior.

Reviewed By: bethebunny

Differential Revision: D16232774

fbshipit-source-id: d8fd642546c6b441e6d5292649bc5193459fd081