Force tensor type of `dense_feat` to be a float.
Summary:
Developing a PyText `DocModel` as a TorchScript model causes the following error to be thrown (despite training and publishing successfully) when calling it:

{P75435136}

This occurs because without explicitly setting the `dtype` of the `dense_feat` tensor, it infers it to be a tensor of doubles (likely because the data arrives via a [`Generic Value`](https://fburl.com/epi2r0x7) union that holds the data as doubles). This then gets concatenated to a separate tensor of floats, and this results in the above type error.

It's not feasible to change this union to accept a float field because it gets used in many other places, so instead this diff forces the tensor to be a tensor of floats, resolving this type error. It's also more robust if we force the tensor to be of the correct type here, as this is what we'll need later on when we concatenate it.

Differential Revision: D16540752

fbshipit-source-id: 80d668e2a2d44e53a03189b9781a68278c1be1ce