implemented bottleneck separable convolutions (#855)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/855

Creates bottleneck layers for separable convolutions. Downsampling, convolving, and then upsampling significantly cuts the number of parameters with minimal loss in performance. This diff is a variant of the traditional bottleneck method -- instead of upsampling directly in the pointwise convolution, we split it into two pieces where the first convolution downsamples into a (sufficiently small) low dimension and the second convolution upsamples into the target (higher) dimension.

Example: Given an input with 256 channels, the depthwise channel spatially convolves it with 256 channels. Then, instead of projecting the channel space to 256 with the pointwise convolution, it is split into two pieces. The first pointwise convolution projects the channel space to 64, then the second pointwise convolution projects the channel space to 256 -- as we intended.

{F172681616}

Reviewed By: geof90

Differential Revision: D16563566

fbshipit-source-id: 1cf310315739802aa5ea6a34efb4c8ab771a3c63