removed lm labels exporting (#866)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/866

Removes caffe2 labels exporting for the language model output layer. Labels overhead adds a lot of latency overhead due to the log softmax operation and loading/splitting scores into the output blobs. This is especially problematic for language models as they have large vocabularies -- performing 8-15k splits/loads can be very expensive. Shreyan has observed that processing the output blobs alone takes ~300ms, which is well over what it takes the model to process the original sequence.

The original motivation for including output layer blobs was that the client would have a copy of the vocabulary. However, it looks like this is unnecessary as the vocabulary can be obtained from the init net using the exporter created in D16268693. The `tokens_vals_index_vocab` blob contains a binarized version of the vocab. Each binarized string can be decoded as utf-8 to recover the original string.

Reviewed By: geof90

Differential Revision: D16617730

fbshipit-source-id: 4d722a7ea307ad88de94840564841a3a522cb4d2