implemented cnn pooling for doc classification (#872)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/872

Implements max and average pooling for DeepCNNRepresentation -- this allows it to be used in tasks where each word in a sequence doesn't necessarily need to have its own representation. Attention is commonly used in NLP to attend to important parts of the sequence, but max and average pooling are simpler (but strong) baselines that can compress sequential representations. The resulting representation can be used in tasks such as document classification, which we perform experiments with.

Differential Revision: D16631634

fbshipit-source-id: 93a1a3e7d9e7cf54ceebf41523134d9e2c3ba6b7