Fix issue with some tensorizers still re-initializing vocab when loaded from saved state (#848)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/848

Not all tensorizers have the

```
if self.vocab:
    return
```
logic in their `initialize()` method. So, some tensorizers will still re-initialize, and load the full training data set, which defeats the purpose of pickling them in the first place.

Expecting every tensorizer author to include that logic in `initialize()` seems very redundant. The platform should be smart enough to support skipping the full pass of training data if the tensorizers were loaded from a saved state. This diff implements that.

Reviewed By: seayoung1112

Differential Revision: D16533827

fbshipit-source-id: 7b05d031521ac35c2c05dbd2db89b437f7e57fc0