move fairseq dependency from master code to current release
Summary:
We cannot release pytext with an editable dependency.

The code dependency was added because we frequently depend on the latest fairseq code.  I suggest we go back to the code dependency after we do the release.  We can coordinate releases with fairseq and follow this procedure in the future.

Differential Revision: D16807150

fbshipit-source-id: d8ad585ef084fa69ad136ee61f636b05690d2b2d