move fairseq dependency from master code to current release (#908)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/908

We cannot release pytext with an editable dependency.

The code dependency was added because we frequently depend on the latest fairseq code.  I suggest we go back to the code dependency after we do the release.  We can coordinate releases with fairseq and follow this procedure in the future.

Reviewed By: Titousensei

Differential Revision: D16807150

fbshipit-source-id: ce997e974a2770235e97b9b73f2f49d9473d109a