EASY [PyText] fix OSS release
Summary:
this is a reconciliation diff to match what was just released to
as v0.2.2. There was several issues during this release, and I uploaded to
pypi a package that was broken. Pypi does not allow overwriting, so I had to
increase the version number.

I manually fixed the source (because it was minor packaging and dependency
problems) and pushed package 0.2.2.  This diff includes those fixes after
the fact to make the sources consistent with the release.

Changes are:
- fix fairseq requirement (separate diff, landing now)
- add __init__.py to make qna a module included in the release
- amend CHANGELOG to match version numbers

Reviewed By: neo315

Differential Revision: D16809591

fbshipit-source-id: ca6ef6598939fbdbb9a618331a84ae14c94c4094