Add `ConsistentXLMModel`
Summary: Add an XLM model that accepts two text columns (and no label columns) as input. The model will evaluate (soft) predictions on the reference text input and treat that as the target distribution for the text in the "tokens" text input. This can be used for example when the two text columns are translations of each other (possibly multi-tasked with a regular labeled task).

Differential Revision: D16786687

fbshipit-source-id: 6976a9bdf488f5e00fcafb4d023a71188b735bb1