Optimize Gelu module for caffe2 export (#918)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/918

TIL ONNX->Caffe2 is very memory inefficient, it creates an intermediate blob for each intermediate output. So, the Gelu operator creates a lot of intermediate ops since it does a bunch of math.

Fix is to use the caffe2 Gelu operator, so all that computation is captured in a single op.

https://pxl.cl/HzGf

Differential Revision: D16849396

fbshipit-source-id: a17908daff58b2c005afbc72fcb8dc46c37d075d