Optimize Gelu module for caffe2 export (#918)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/918

TIL ONNX->Caffe2 is very memory inefficient, it creates an intermediate blob for each intermediate output. So, the Gelu operator creates a lot of intermediate ops since it does a bunch of math.

Fix is to use the caffe2 Gelu operator, so all that computation is captured in a single op.

https://pxl.cl/HzGf

Reviewed By: seayoung1112

Differential Revision: D16849396

fbshipit-source-id: 40920b74caa3ad4244f2af5e24f11d4f123b935f