Add `ConsistentXLMModel` (#913)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/913

Add an XLM model that accepts two text columns (and no label columns) as input. The model will evaluate (soft) predictions on the reference text input and treat that as the target distribution for the text in the "tokens" text input. This can be used for example when the two text columns are translations of each other (possibly multi-tasked with a regular labeled task).

Reviewed By: rutyrinott

Differential Revision: D16786687

fbshipit-source-id: a62d3ec4c27cc7e891375d38459ba1621e5b9d9d