Make PerLabelConfusions output in a deterministic order.
Summary:
When outputting the per label confusions, the `compute_metrics` method iterates through an unordered dictionary, this makes it non-deterministic which order these labels are printed. This becomes a problem when we want to consistently refer to an output value, for instance when setting the objective_attribute in a param sweep where we might use an argument like `outputs.doc_results_pair.test.per_label_prf1_scores[X].f1` to refer to a particular index `X` of the output labels.

This diff sorts the labels before computing the stats, so that they will always appear in the same order for consecutive runs.

Differential Revision: D16915421

fbshipit-source-id: 7f2c08a1a4f98fd2964ece13efacf1e2c5c1bba5