Make PerLabelConfusions output in a deterministic order. (#923)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/923

When outputting the per label confusions, the `compute_metrics` method iterates through an unordered dictionary, this makes it non-deterministic which order these labels are printed. This becomes a problem when we want to consistently refer to an output value, for instance when setting the objective_attribute in a param sweep where we might use an argument like `outputs.doc_results_pair.test.per_label_prf1_scores[X].f1` to refer to a particular index `X` of the output labels.

This diff sorts the labels before computing the stats, so that they will always appear in the same order for consecutive runs.

Reviewed By: rutyrinott

Differential Revision: D16915421

fbshipit-source-id: 327c5d1d78d85a743e313334d595375816ab69c4