Perform a two-layered pooling in PoolingBatcher (#949)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/949

`PoolingBatcher` will now shuffle multiple pools' worth of data, before splitting that into pools which are then sorted, and split into batches. The size of this "super-pool" is user-selectable via `num_shuffled_pools` (set to 1 by default to mimic the old behaviour).

By having a large `num_shuffled_pools` and `pool_num_batches=1`, data is shuffled and it is only sorted within the batch. This leads to potentially lower performance (because input lengths inside a batch may be quite varied), but a better approximation of the true gradient. On the other hand, a large `pool_num_batches` means that inputs within a batch have more similar lengths, which is more efficient.

Reviewed By: ArmenAg

Differential Revision: D17185771

fbshipit-source-id: 82a4976d21088a354e822db40ef01002649eb6a2