config implicit class should be the default value (#951)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/951

DocumentClassificationTask.Config declares

    model: BaseModel.Config = DocModel.Config()

When a config references "model" without specifying the class name (like in the example below), we expect it to use DocModel.

  {
    "task": {
      "DocumentClassificationTask": {
        "model": {
          "representation": {
            "DocNNRepresentation": {}
          }
        }
    ...

But what really happens is it uses BaseModel and fails with the error message:

  pytext.config.serialize.ConfigParseError: Unknown fields for class BaseModel.Config with fields {'inputs'}
  detected in config json: {'representation'}"

If your json does not reference "model", it works because in this case we use the default value.

This diff fixes this problem by always using the default value when present, whether the parameter name is in the config or not.

Differential Revision: D17208598

fbshipit-source-id: 8a0119c26da4271d3581ea85e91d7a73f46fa5a1