Fixes loading model snapshot for incremental training
Summary:
When we load snapshot in PyText, we take serialized task from it, and ignore task passed in the current config, which doesn't seem to be correct. The behavior I want is to load a model from the snapshot, and continue training according to the task specified in the current config (not serialized one).

This diff takes tensorizers and model from loaded snapshot, and initializes the current task with it.

Differential Revision: D17195794

fbshipit-source-id: 4b8aa5985d56e84b5485ba7a35693cb18ce3cfc0