Fix bug in copy_unk (#964)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/964

When the copy_unk flag is set to true. Any unk that is produced in the output of the Seq2Seq model is replaced by the token that was mapped to unk from the utterance. This is a easy way to get gains since outputs with unk are always wrong.

Looking at the old code for copying the unk token we see that TorchScript optimizes out the actual search of the unk token in utterance:

{F207887831}

This diff updates the code to produce the correct TorchScript Graph

{F207888470}

Reviewed By: arbabu123

Differential Revision: D17213086

fbshipit-source-id: 9f02f7dddf79671e96feb28b0581ded504cb69fc