Drop rows with insufficient columns in TSV data source (#954)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/954

We use `csv.DictReader` when reading in TSV files which when reads rows with insufficient number of columns, sets the trailing columns to None. This diff introduces a flag to ignore such rows. If not ignored, a tensorizer could throw KeyError in `numberize()` method since `None` is not a valid key.

Reviewed By: borguz

Differential Revision: D17223777

fbshipit-source-id: de21f5fe8ed4f2e06032f82f522e6539d6c48c3f