Drop rows with insufficient columns in TSV data source (#954)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/954

We use `csv.DictReader` when reading in TSV files which when reads rows with insufficient number of columns, sets the trailing columns to None. This diff introduces a flag to ignore such rows. If not ignored, a tensorizer could throw KeyError in `numberize()` method since `None` is not a valid key.

Reviewed By: borguz

Differential Revision: D17223777

fbshipit-source-id: 1024483d88b2b2913050f439a4f1af6abdc1de3a