Add dense feature normalization to Char-LSTM TorchScript model.
Summary:
In D16357113 we added dense feature tokenisation to the `DocModel` architecture. As the `ByteTokensDocumentModel` extends the `DocModel` class and uses the `FloatListTensorizer` class, feature normalisation was already supported at training time after this diff was landed, however it was not supported at production time because the normalisation was not performed during the `torchscriptify` forward function. This presents a problem because the model can be trained on normalized data, but then won't be able to normalize fresh data at inference time, producing unusual results.

This diff adds normalisation in the `torchscriptify` forward function so that it can be used at inference time on Char-LSTM models.

Just like with the `DocModel`, if `normalize` is set to `false` in the config, then the normalization method just becomes the identity function, this is useful because we don't need to add any extra control flow directly in the forward function.

Reviewed By: snisarg

Differential Revision: D17358479

fbshipit-source-id: 4a92444963f52290d538377070bf872c9b42aba9