handle long documents in squad qa datasource and models (#975)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/975

BERT encoder has a maximum sequence length, which means long paragraphs cat get cut off when training / evaluating QA models.  In the original Squad dataset, it's very rare to have a paragraph and an answer span that will get truncated, however it can be more common in other datasets.

This diff implements chunking for long paragraphs such that:
- each chunk is fixed size in terms of character length
- a minimum overlap can be specified as a fraction of chunk size

In metric reporter, we aggregate by sample id and return the highest scoring span.

To share chunking logic between SquadDataSource and SquadTSVDataSource, I merged them into one class, with the added benefit that we can now train on a .json file and test on a tsv file and visa versa.

Reviewed By: hikushalhere

Differential Revision: D17350243

fbshipit-source-id: 77c56b7c1a6160ffb38231e822a1f8a18845a53f