Remove vocab from cuda (#955)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/955

We have users who can't train models on extremely large embeddings because we try to allocate space for that on the GPU.

With this diff, in training, we add a flag which users can set explicitly to keep the embedding layer on CPU even when the model is getting trained on GPUs. This is not default because we need the user to know that there will be a cost associated moving the tensors on and off the GPU.

Note that this only applies during training.

Also note that this does not work in a multi-GPU environment because of the way the weights are synced via NCCL.

Reviewed By: chenyangyu1988

Differential Revision: D17114398

fbshipit-source-id: ee4142fb9cada502b02219f2a55e6afaf122d4c1