Implement Cyclic CopyUnk Functionality (#1063)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1063

This implementation implements cyclic copying of unk tokens for when number of unk in utterances and target is the same, as well as copy unk when there's only one unk. For all the other cases we implement cyclic copying. This implementation is backward compatible with expected behavior and latency is not changed.

Differential Revision: D18032736

fbshipit-source-id: be09981b4452ce773b3f7c0ab80d21952c12c611