Enable passing `model_test_params` to model during pytext testing
Summary:
Enable configuring test params during pytext testing and configuring a list of test params to the TrainAndMultiTestWorkflow and RunMultilingualKDWorkflow. These params are passed to the model to prepare for testing.

In particular, for stacked transformers we train a single model, but want to test each K-layer model individually. So we need to configure which layer to test for each test run.

Differential Revision: D18097006

fbshipit-source-id: e169abb59bea7f7c64d2c4e6b0be8dceb9e6db29