Fix label_weights in DocModel (#1081)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1081

While playing around with the label_weights for OOD found a bug in the way the loss is constructed for the DocModel class resulting in the label_weights not being passed to the loss constructor. Verified in a bento that the trained model had empty label_weights in the loss. This fixes the bug and properly passes the weights to the constructor

Differential Revision: D18176357

fbshipit-source-id: abb5a9bade411b83240d0bb0c5a7d8aa09dc234f