Standardize RoBERTa Tensorizer Vocab Creation
Summary: As part of the Tensorizer refactor, I standarize vocab creation for the RoBERTa tesorizer i.e. remove vocab creation from the tokenizer and bring it into the tensorizer. I also make the special tokens for RoBERTa confiurable so that we don't need a separate tensorizer if we decide to train RoBERTa with different special tokens.

Differential Revision: D18289234

fbshipit-source-id: 0eeacebf3ecd6881aefc02bd70a19db5127509c5