Standardize RoBERTa Tensorizer Vocab Creation (#1113)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1113

As part of the Tensorizer refactor, I standarize vocab creation for the RoBERTa tesorizer i.e. remove vocab creation from the tokenizer and bring it into the tensorizer. I also make the special tokens for RoBERTa confiurable so that we don't need a separate tensorizer if we decide to train RoBERTa with different special tokens. I also revert some of the changes made in D17974656 which break loading of fairseq vocab for all tensorizers.

Reviewed By: chenyangyu1988

Differential Revision: D18289234

fbshipit-source-id: 6b8cbb911e3942310edbde2b7d2dac1a4c18235d