Adding Google SentencePiece as a Tokenizer (#1106)
Summary:
This diff adds SentencePiece as a pip requirement, and a tokenizer shell
for PyText

## Motivation and Context

We need SentencePiece to support modern cross lingual models

## How Has This Been Tested

A unit test has been added.

## Types of changes

- [ ] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [x] I have completed my CLA (see **CONTRIBUTING**)
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1106

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

Tests are non-issue because the TARGET file was missing dependency, which cannot be filled up in OSS. Look at the diff stacked on top of this one for tests.

Reviewed By: hudeven

Differential Revision: D18309626

Pulled By: snisarg

fbshipit-source-id: a0d16417023e237eb29a9355c22e203e380efbb5