RoBERTa-QA JIT (#1088)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1088

This diff torchscriptifies RoBERTa-QA to make it exportable in the same manner as BERT-QA.
It also converts the has_answer_logits tensor to float before doing an argmax in squad_output_layer in order to use the FP16FairseqOptimizer. This otherwise breaks as argmax_cuda isn't supported for "Half" datatype.

Reviewed By: hikushalhere

Differential Revision: D18218091

fbshipit-source-id: 007ed1df6422efab74ee061baa05884c24200e42