Complete training loop gracefully even if no timing is reported
Summary: PyText always requires/assums that at least one "timing" exists at the end of the training and tries to print it as an ascii table. However, if one simply does not have any `timing.time("foo")` call during the training, the entire training loop crashes in the end, which is I don't think desirable behavior and not all people might not be that interested in timing info at the early phase iteration/research.

Differential Revision: D18461987

fbshipit-source-id: ed3e2918ced9f12d172c489cf9ef3b0237729aca