Explicitly set trivial start and end losses for MRC models to float type
Summary:
For MRC models, when true answer is not present, start and end position losses are set to 0. PyTorch complains that the value of tensors that require grads must be floating point dtype. Default behavior of `torch.tensor()` constructor is infer data type from data if dtype is not mentioned.

Note: I couldn't find anything that says the default behavior has changed though. Not sure why we didn't encounter this before.

Differential Revision: D18450671

fbshipit-source-id: a27adb8934ef40cd0d624f6cb4a4b304501c9fac