Complete training loop gracefully even if no timing is reported (#1128)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1128

PyText always requires/assums that at least one "timing" exists at the end of the training and tries to print it as an ascii table. However, if one simply does not have any `timing.time("foo")` call during the training, the entire training loop crashes in the end, which is I don't think desirable behavior and not all people might not be that interested in timing info at the early phase iteration/research.

Reviewed By: kmalik22

Differential Revision: D18461987

fbshipit-source-id: 0f8415c1777072f0e2776c7d331991c167c187b4