Proper behavior for do_eval=False
Summary:
In trainer, the option `do_eval` (https://fburl.com/diffusion/lopy7hig) can be set to `False` to skip eval stage in training. However, the eval stage is always run in the finalize step (see changes in this diff). This can cause issues when eval data is not provided, see the failed workflow for example:
f151938461

In this diff, if `do_eval` is set to `False`, we also skip the eval operation in the finalize step.

Differential Revision: D18590815

fbshipit-source-id: 7041e95e709cfa3873db6f165a76a4d0d0160d87