Proper training behavior if setting do_eval=False (#1155)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1155

In trainer, the option `do_eval` (https://fburl.com/diffusion/lopy7hig) can be set to False to skip eval stage in training. However, in the `finalize` step the eval stage is always run (see changes in this diff). This can cause issues when eval data is empty, and `do_eval` is set to False, for example see this failed workflow:
f151938461

In this diff, we also skip eval in the `finalize` step if `do_eval` is set to False.

Reviewed By: seayoung1112

Differential Revision: D18590815

fbshipit-source-id: 9e079fabbcda97122bd868bcd89f6ffd6bfc615c