Fix index error in dict embedding when exported to Caffe2 (#1182)
Summary:
`feats[feats == self.unk_index] = self.pad_index` gets converted to `IndexPut` operator in caffe2, which does not work on multidimensional tensors. In this case, `feats` is a 2d array of `batch_size, dict_len * seq_len`, so we get an IndexError in the exported model.

This diff fixes it by generating an equal sized scalar filled tensor first and then using the `Where` operator to update values.

Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1182

Reviewed By: seayoung1112

Differential Revision: D18805595

fbshipit-source-id: 7b32af817c6db7a6fa841c3eccea2b2f4ac74515