Mobile compatible version of Dict Embedding (#1230)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1230

`torch.where` has some problem with ONNX in mobile, where it is not supported. Since our on device model relies on dictinoary embeddings this stops us from launching an update.

This diff adds a mobile compatible version of `torch.where` expressed as element wise matrix operations, and extends the config to determine whether the mobile version should be used or not.

Some notes:

 - It is `(1 - mask.long())` instead of `torch.logical_not(mask)` because `torch.logical_not` is not supported in mobile.
 - Discussion here: https://fb.workplace.com/groups/1405155842844877/permalink/2672442776116171/

Reviewed By: geof90

Differential Revision: D19377525

fbshipit-source-id: d4023105512d640648725bc00ebfa2f315596970