Fix KD in fluent2 (#1203)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1203

KD in fluent2 was broken because the main (student) model always expects a list of strings (tokens). However, when reading the tokens field from a teacher-generated file, it instead receives a string. Modify the student token tensorizer to parse the string input as JSON into a list of strings (as for non-KD).

Also make the test more robust (the test checks that training examples are classified correctly after training, so remove dropout and set high lr).

Reviewed By: rutyrinott

Differential Revision: D18822211

fbshipit-source-id: 4555940b300cbb6b56018991512bdd4fcf4fcb6e