Add support for returning start & end indices.
Summary:
The SentencePieceTokenizer currently returns -1s as start and end indices. This diff adds support for returning the correct start and end indices which would be useful for some word tagging and span extraction models.

SPM tokenizes the first piece of every word by prefixing a separator character "\u2581" which looks like "▁". For example, see the expected tokens in the attached unit test.
We need to strip this character off to get the original token in order to locate it and compute the correct start and end index.

Reviewed By: chenyangyu1988

Differential Revision: D19626317

fbshipit-source-id: 7254fc43bc0064ee7d2e6726cfb10d96a2b1b192